Move timeEntry to features
