+++ package.json
@@ -6,7 +6,7 @@
   "license": "MIT",
   "repository": {},
   "scripts": {
-    "start": "ts-node-dev --respawn --transpileOnly ./src/server/index.ts start",
+    "start": "cross-env NODE_PATH=./src ts-node-dev -P ./tsconfig.json --respawn --transpileOnly ./src/server/index.ts start",
     "build": "cross-env NODE_ENV=production webpack",
     "test": "jest",
     "lint:code": "tslint --project .",
