--- src/api/deleteTiming/contract.ts
@@ -1,8 +1,0 @@
-import { RequestMetadata } from '../../shared/http';
-
-export const formatUrl = (id: string) => `/api/time/${id}`;
-
-export const metadata: RequestMetadata = {
-    method: 'DELETE',
-    url: '/api/time/:id',
-};
