+++ src/features/timeEntry/components/issueFilter/index.scss
