+++ src/features/timeEntry/components/issueList/index.scss
