+++ src/store/createStore.ts
@@ -1,8 +1,8 @@
 import { createStore, applyMiddleware, combineReducers } from 'redux';
 import { reducer as activityListReducer } from './activityList/reducer';
 import { reducer as timingsPageReducer }  from './timingList/reducer';
-import { reducer as timingsFormReducer }  from './timing/reducer';
-import { reducer as issuesReducer } from './issues/reducer';
+import { reducer as timingsFormReducer }  from '../features/timeEntry/state/timing/reducer';
+import { reducer as issuesReducer } from 'features/timeEntry/state/issues/reducer';
 import { reducer as enumerationsReducer } from './enumerations/reducer';
 import { reducer as authReducer } from '../features/auth/state/reducer';
 import { reducer as globalReducer } from './@global/reducer';
